requirements.txt
numpy
torch
transformers
Pillow
tqdm
streamlit
scikit-learn
pycocotools
accelerate 
bitsandbytes